+++ src/constants/index.js
@@ -1,0 +1,3 @@
+export const VALID_NUMBER_LENGTH = 3;
+export const MIN_RANDOM_NUMBER = 1;
+export const MAX_RANDOM_NUMBER = 9;
